Move classes to separate packages
