+++ path-tracker-pipelines/pom.xml
@@ -29,8 +29,8 @@ under the License.
 	<artifactId>path-tracker-pipelines</artifactId>
 
 	<properties>
-		<maven.compiler.source>11</maven.compiler.source>
-		<maven.compiler.target>11</maven.compiler.target>
+		<maven.compiler.source>8</maven.compiler.source>
+		<maven.compiler.target>8</maven.compiler.target>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<flink.version>1.20-SNAPSHOT</flink.version>
 		<slf4j.version>1.7.26</slf4j.version>
