+++ path-tracker-pipelines/src/main/java/org/example/operator/DecorateRichFilterFunction.java
@@ -1,0 +1,12 @@
+package org.example.operator;
+
+import org.apache.flink.api.common.functions.FilterFunction;
+
+import org.example.datasource.DecorateRecord;
+
+public class DecorateRichFilterFunction<IN> extends BaseDecorateRichFunction implements FilterFunction<DecorateRecord<IN>> {
+    @Override
+    public boolean filter(DecorateRecord<IN> inDecorateRecord) throws Exception {
+        return false;
+    }
+}
