+++ path-tracker-pipelines/src/main/java/org/example/pipelines/GlobalSortPipeline.java
@@ -1,0 +1,4 @@
+package org.example.pipelines;
+
+public class GlobalSortPipeline {
+}
